WebSailor/nanoWebSailor/requirements.txt
# WebSailor项目依赖包

# 核心依赖
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0

# 数据处理
pandas>=1.5.0
numpy>=1.24.0
jsonlines>=3.1.0
pydantic>=2.0.0

# Web工具
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
jina>=3.18.0

# 知识图谱和数据库
SPARQLWrapper>=2.0.0
networkx>=3.1.0
rdflib>=6.3.0
scipy==1.14.0

# 评估和指标
# scikit-learn>=1.3.0
rouge-score>=0.1.2
bert-score>=0.3.13

# 可视化和日志
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.15.0
loguru>=0.7.0
plotly>=5.18.0

# 配置管理
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0.0

# 异步和并发
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

# 工具类
tqdm>=4.65.0
rich>=13.0.0
typer>=0.9.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0 

openai>=1.0.0 